system: Move conf file read functionality to separate implementation file
Separating conf file read functionality allows the majority of conf
functinality to be usable where stdlib fopen/fgets/fopen isn't available.
Also removed spdk_fopen/spdk_fgets/spdk_fclose as they are unnecessary
abstractions now.
