+++ app/vhost/Makefile
@@ -42,8 +42,8 @@ CFLAGS += $(ENV_CFLAGS)
 
 C_SRCS := vhost.c
 
-SPDK_LIB_LIST = jsonrpc json rpc bdev_rpc bdev scsi net copy trace conf
-SPDK_LIB_LIST += util log log_rpc event app_rpc
+SPDK_LIB_LIST = jsonrpc json rpc bdev_rpc bdev scsi net copy trace
+SPDK_LIB_LIST += util log log_rpc event app_rpc conf
 SPDK_LIB_LIST += vhost rte_vhost
 
 LIBS += $(BLOCKDEV_MODULES_LINKER_ARGS) \
